rename: exception 패키지를 common 모듈로 이동
