+++ packy-api/src/main/java/com/dilly/auth/application/AuthService.java
@@ -18,9 +18,9 @@ import com.dilly.auth.dto.response.SignInResponse;
 import com.dilly.auth.model.AppleAccountInfo;
 import com.dilly.auth.model.AppleToken;
 import com.dilly.auth.model.KakaoResource;
+import com.dilly.exception.ErrorCode;
+import com.dilly.exception.UnsupportedException;
 import com.dilly.gift.ProfileImage;
-import com.dilly.global.exception.UnsupportedException;
-import com.dilly.global.response.ErrorCode;
 import com.dilly.global.utils.SecurityUtil;
 import com.dilly.jwt.JwtService;
 import com.dilly.jwt.RefreshToken;
