+++ packy-api/src/main/java/com/dilly/gift/domain/BoxReader.java
@@ -5,10 +5,10 @@ import java.util.List;
 import org.springframework.stereotype.Component;
 
 import com.dilly.admin.dto.response.BoxImgResponse;
+import com.dilly.exception.ErrorCode;
+import com.dilly.exception.entitynotfound.EntityNotFoundException;
 import com.dilly.gift.Box;
 import com.dilly.gift.dao.BoxRepository;
-import com.dilly.global.exception.entitynotfound.EntityNotFoundException;
-import com.dilly.global.response.ErrorCode;
 
 import lombok.RequiredArgsConstructor;
