--- packy-api/src/main/java/com/dilly/global/exception/AuthorizationFailedException.java
@@ -1,10 +1,0 @@
-package com.dilly.global.exception;
-
-import com.dilly.global.response.ErrorCode;
-
-public class AuthorizationFailedException extends BusinessException {
-
-    public AuthorizationFailedException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
