--- packy-api/src/main/java/com/dilly/global/exception/BusinessException.java
@@ -1,15 +1,0 @@
-package com.dilly.global.exception;
-
-import com.dilly.global.response.ErrorCode;
-import lombok.Getter;
-
-@Getter
-public class BusinessException extends RuntimeException {
-
-    private final ErrorCode errorCode;
-
-    public BusinessException(ErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-    }
-}
