+++ packy-api/src/main/java/com/dilly/global/response/DataResponseDto.java
@@ -1,5 +1,7 @@
 package com.dilly.global.response;
 
+import com.dilly.exception.ErrorCode;
+
 import lombok.Getter;
 
 @Getter
