+++ packy-api/src/main/java/com/dilly/global/response/ResponseDto.java
@@ -1,5 +1,7 @@
 package com.dilly.global.response;
 
+import com.dilly.exception.ErrorCode;
+
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
