+++ packy-api/src/main/java/com/dilly/global/utils/SecurityUtil.java
@@ -3,8 +3,8 @@ package com.dilly.global.utils;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 
-import com.dilly.global.exception.AuthorizationFailedException;
-import com.dilly.global.response.ErrorCode;
+import com.dilly.exception.AuthorizationFailedException;
+import com.dilly.exception.ErrorCode;
 
 import lombok.extern.slf4j.Slf4j;
