+++ packy-common/src/main/java/com/dilly/exception/AuthorizationFailedException.java
@@ -1,0 +1,8 @@
+package com.dilly.exception;
+
+public class AuthorizationFailedException extends BusinessException {
+
+	public AuthorizationFailedException(ErrorCode errorCode) {
+		super(errorCode);
+	}
+}
