+++ packy-common/src/main/java/com/dilly/exception/BusinessException.java
@@ -1,0 +1,14 @@
+package com.dilly.exception;
+
+import lombok.Getter;
+
+@Getter
+public class BusinessException extends RuntimeException {
+
+	private final ErrorCode errorCode;
+
+	public BusinessException(ErrorCode errorCode) {
+		super(errorCode.getMessage());
+		this.errorCode = errorCode;
+	}
+}
