+++ packy-common/src/main/java/com/dilly/exception/alreadyexist/AlreadyExistException.java
@@ -1,0 +1,11 @@
+package com.dilly.exception.alreadyexist;
+
+import com.dilly.exception.BusinessException;
+import com.dilly.exception.ErrorCode;
+
+public class AlreadyExistException extends BusinessException {
+
+	public AlreadyExistException(ErrorCode errorCode) {
+		super(errorCode);
+	}
+}
