+++ packy-common/src/main/java/com/dilly/exception/internalserver/AppleServerException.java
@@ -1,8 +1,8 @@
-package com.dilly.global.exception.internalserver;
+package com.dilly.exception.internalserver;
 
-import static com.dilly.global.response.ErrorCode.*;
+import static com.dilly.exception.ErrorCode.*;
 
-import com.dilly.global.response.ErrorCode;
+import com.dilly.exception.ErrorCode;
 
 public class AppleServerException extends InternalServerException {
