+++ packy-common/src/main/java/com/dilly/exception/internalserver/InternalServerException.java
@@ -1,7 +1,7 @@
-package com.dilly.global.exception.internalserver;
+package com.dilly.exception.internalserver;
 
-import com.dilly.global.exception.BusinessException;
-import com.dilly.global.response.ErrorCode;
+import com.dilly.exception.BusinessException;
+import com.dilly.exception.ErrorCode;
 
 public class InternalServerException extends BusinessException {
